refactor(now-playing): extract renderer and context methods into nowPlayingListRenderer.ts
Moves all shared rendering, layout, and list-refresh logic out of the
monolithic NowPlayingCommand class into a dedicated module-level exported
functions file (src/commands/now-playing/nowPlayingListRenderer.ts).

This is step 5 of the issue-716 extraction plan. Functions extracted:
- buildNowPlayingListLines, buildNowPlayingListContainer,
  buildNowPlayingMessageContainer, buildComponentPayload
- buildNowPlayingAttachments, buildNowPlayingListPayload,
  buildNowPlayingJournalSelectRow, buildNowPlayingCompositeImageUrl,
  buildNowPlayingCompositeSourceHash
- buildNowPlayingActionRow, buildNowPlayingManageRow,
  buildNowPlayingEditMenuComponents, returnToNowPlayingEditMenu,
  buildNowPlayingEditInitialComponents, withPmNowPlayingList
- withNowPlayingActions, hasDisplayableNowPlayingNotes,
  refreshNowPlayingListFromContext, getNowPlayingShowNotesState,
  buildNowPlayingEntryComponents, trimTextDisplayContent,
  buildNowPlayingMemberSelect
